fv_enc/requirements.txt
cocotb
cocotb-bus
cocotbext-axi
Cython
exceptiongroup
find-libpython
iniconfig
numpy
